script for providing owl-liquidity
